+++ CHANGELOG.md
@@ -18,6 +18,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [0.1.0]
 ### Added
+- ONTOLOGY-28 Added view export as csv (zipped file containing basic and full CSV data)
 - ONTOLOGY-28 Added view export as pdf
 - ONTOLOGY-28 Added view export as png and jpeg
 - ONTOLOGY-36 Added settings to sidebar